+++ CHANGES.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Add shell-style escaping support for the System Runner
+
 ## 0.5.0 - 2024-06-27
 
 ### Added